Upgrading the carbon commons version and changing SPARK_EVENTING_TASK…

+++ components/analytics-processors/org.wso2.carbon.analytics.spark.event/src/main/java/org/wso2/carbon/analytics/spark/event/EventingConstants.java
@@ -31,7 +31,7 @@ public class EventingConstants {
     public static final String ANALYTICS_SPARK_EVENTING_TASK_NAME = "STORE_EVENT_ROUTER_TASK";
     public static final String DISABLE_EVENT_SINK_SYS_PROP = "disableEventSink";
     public static final String DISABLE_SPARK_EVENTING_TASK_SYS_PROP = "disableSparkEventingTask";
-    public static final int SPARK_EVENTING_TASK_RUN_INTERVAL_MS = 10000;
+    public static final long SPARK_EVENTING_TASK_RUN_INTERVAL_MS = 10000;
     public static final String EVENT_META_DATA_PREFIX = "meta_";
     public static final String EVENT_CORRELATION_DATA_PREFIX = "correlation_";
